Crate parse_hyperlinks

Library and application for parsing hyperlinks and link reference definitions in Markdown, reStructuredText, Asciidoc and HTML format. The library implements the CommonMark Specification 0.29, reStructuredText Markup Specification (revision 8571, date 2020-10-28), Asciidoctor User Manual, chapter 26 (date 2020-12-03) and the HTML 5.2: section 4.5. specification.

Modules

iterator

Module providing an iterator over the hyperlinks found in the input text. Consult the documentation of parser::take_link() to see a list of supported markup languages. The iterator resolves link references.

parser

This module implements parsers to extract hyperlinks and link reference definitions from text input.

renderer

A set of functions providing markup source code to HTML renderer, that make hyperlinks clickable.

Functions

take_until_unbalanced

A parser similar to nom::bytes::complete::take_until(), but that does not stop at balanced opening and closing tags. It is designed to work inside the nom::sequence::delimited() parser.