Module parse_hyperlinks::iterator

Module providing an iterator over the hyperlinks found in the input text. Consult the documentation of parser::take_link() to see a list of supported markup languages. The iterator resolves link references.

Structs

Hyperlink

A struct to hold the iterator’s state and a pointer to the input text. The iterator’s next() method returns a tuple with 2 tuples inside:

Functions

first_hyperlink

Recognizes hyperlinks in all supported markup languages and returns the first hyperlink found as tuple: Some((link_text, link_destination, link_title)).