Module parse_hyperlinks::parser

This module implements parsers to extract hyperlinks and link reference definitions from text input.

Modules

asciidoc

This module implements parsers for Asciidoc hyperlinks.

html

This module implements parsers for HTML hyperlinks.

markdown

This module implements parsers for Markdown hyperlinks.

restructured_text

This module implements parsers for RestructuredText hyperlinks.

Enums

Link

A link can be an inline link, a reference link, a link reference definition, a combined inline link / link reference definition or a reference alias. This is the main return type of this API.

Constants

LABEL_LEN_MAX

Link max label. This limits the damage of a forgotten closing brackets. CommonMark Spec

Functions

take_link

Consumes the input until it finds a Markdown, RestructuredText, Asciidoc or HTML formatted inline link (Text2Dest), reference link (Text2Label), link reference definition (Label2Dest) or reference alias (Label2Label).

take_text2dest_label2dest

Consumes the input until it finds a Markdown, RestructuredText, Asciidoc or HTML formatted inline link (Text2Dest) or link reference definition (Label2Dest).