Module parse_hyperlinks::parser::parse

This module implements parsers to extract hyperlinks and link reference definitions from text input. The parsers search for Markdown, ReStructuredText, Asciidoc, Wikitext and HTML formatted links.

Constants

LABEL_LEN_MAX

Link max label. This limits the damage of a forgotten closing brackets. CommonMark Spec

Functions

take_link

Consumes the input until it finds a Markdown, RestructuredText, Asciidoc or HTML formatted inline link (Text2Dest), reference link (Text2Label), link reference definition (Label2Dest) or reference alias (Label2Label).

take_text2dest_label2dest

Consumes the input until it finds a Markdown, RestructuredText, Asciidoc or HTML formatted inline link (Text2Dest) or link reference definition (Label2Dest).