Module parse_hyperlinks::parser

This module implements parsers to extract hyperlinks and link reference definitions from text input.

Modules

asciidoc

This module implements parsers for Asciidoc hyperlinks.

html

This module implements parsers for HTML hyperlinks.

markdown

This module implements parsers for Markdown hyperlinks.

parse

This module implements parsers to extract hyperlinks and link reference definitions from text input. The parsers search for Markdown, ReStructuredText, Asciidoc, Wikitext and HTML formatted links.

restructured_text

This module implements parsers for RestructuredText hyperlinks.

wikitext

This module implements parsers for Wikitext hyperlinks.

Enums

Link

A link can be an inline link, a reference link, a link reference definition, a combined inline link / link reference definition, a reference alias or an inline image. This is the main return type of this API.